chore: major app restructure
