--- .eslintrc.json
@@ -1,23 +1,0 @@
-{
-	"env": {
-		"browser": true,
-		"es2021": true,
-		"node": true
-	},
-	"extends": [
-		"plugin:@typescript-eslint/recommended",
-		"plugin:@next/next/recommended",
-		"plugin:prettier/recommended"
-	],
-	"overrides": [],
-	"parser": "@typescript-eslint/parser",
-	"parserOptions": {
-		"ecmaVersion": "latest",
-		"sourceType": "module"
-	},
-	"plugins": ["react", "prettier", "@typescript-eslint"],
-	"rules": {
-		"prettier/prettier": ["error"],
-		"react/prop-types": "off"
-	}
-}
