--- kirimase.config.json
@@ -1,18 +1,0 @@
-{
-  "hasSrc": true,
-  "packages": [
-    "drizzle",
-    "trpc",
-    "clerk",
-    "shadcn-ui"
-  ],
-  "preferredPackageManager": "pnpm",
-  "t3": true,
-  "alias": "@",
-  "rootPath": "src/",
-  "orm": "drizzle",
-  "auth": "clerk",
-  "componentLib": "shadcn-ui",
-  "provider": "node-postgres",
-  "driver": "pg"
-}
