+++ package.json
@@ -14,50 +14,24 @@
     "test-ct": "playwright test -c playwright-ct.config",
     "prepare": "is-ci || husky install",
     "prettier": "prettier --write .",
-    "setup-husky": "npx husky install && rm -rf .git/hooks && ln -s ../.husky .git/hooks",
-    "postinstall": "pnpm generate",
-    "db:generate": "drizzle-kit generate:pg",
-    "db:migrate": "tsx src/lib/db/migrate.ts",
-    "db:drop": "drizzle-kit drop",
-    "db:pull": "drizzle-kit introspect:pg",
-    "db:studio": "drizzle-kit studio",
-    "db:check": "drizzle-kit check:pg"
+    "setup-husky": "npx husky install && rm -rf .git/hooks && ln -s ../.husky .git/hooks"
   },
   "dependencies": {
-    "@clerk/nextjs": "^4.27.2",
     "@radix-ui/react-avatar": "^1.0.4",
     "@radix-ui/react-dropdown-menu": "^2.0.6",
     "@radix-ui/react-label": "^2.0.2",
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-toast": "^1.1.5",
-    "@tanstack/react-query": "^4.29.19",
-    "@trpc/client": "^10.44.1",
-    "@trpc/react-query": "^10.44.1",
-    "@trpc/server": "^10.44.1",
     "allotment": "^1.19.3",
     "axios": "^1.6.2",
-    "class-variance-authority": "^0.7.0",
-    "codemirror": "^5.65.13",
     "csp-header": "^5.2.1",
     "dayjs": "^1.11.10",
-    "drizzle-orm": "^0.29.0",
-    "drizzle-zod": "^0.5.1",
     "framer-motion": "^10.16.5",
     "jotai": "^2.5.1",
     "lucide-react": "^0.293.0",
     "next": "^14.0.3",
     "next-themes": "^0.2.1",
-    "postgres": "^3.4.3",
     "react": "18.2.0",
-    "react-beautiful-dnd": "^13.1.1",
-    "react-codemirror2": "^7.2.1",
-    "react-dom": "18.2.0",
-    "react-feather": "^2.0.10",
-    "react-markdown": "^9.0.1",
-    "rehype-sanitize": "^6.0.0",
-    "remark-gfm": "^4.0.0",
-    "superjson": "2.2.1",
-    "tailwindcss-animate": "^1.0.7",
     "uuid": "^9.0.1",
     "zod": "^3.22.4"
   },
@@ -67,18 +41,14 @@
     "@commitlint/prompt-cli": "^18.4.3",
     "@faker-js/faker": "^8.3.1",
     "@playwright/test": "^1.40.0",
-    "@types/codemirror": "^5.60.15",
     "@types/eslint": "^8.44.7",
     "@types/node": "^20.10.0",
     "@types/prettier": "^3.0.0",
     "@types/react": "^18.2.38",
-    "@types/react-beautiful-dnd": "^13.1.7",
     "@types/react-dom": "^18.2.17",
     "@types/uuid": "^9.0.7",
     "@typescript-eslint/eslint-plugin": "^6.12.0",
     "@typescript-eslint/parser": "^6.12.0",
-    "autoprefixer": "^10.4.16",
-    "clsx": "^2.0.0",
     "dotenv": "^16.3.1",
     "drizzle-kit": "^0.20.6",
     "eslint": "^8.54.0",
@@ -87,14 +57,9 @@
     "eslint-plugin-prettier": "^5.0.1",
     "eslint-plugin-react": "^7.33.2",
     "husky": "^8.0.3",
-    "is-ci": "^3.0.1",
     "lint-staged": "^15.1.0",
-    "pg": "^8.11.3",
-    "postcss": "^8.4.31",
     "prettier": "^3.1.0",
     "prettier-plugin-tailwindcss": "^0.5.7",
-    "tailwind-merge": "^2.0.0",
-    "tailwindcss": "^3.3.5",
     "tsx": "^4.6.0",
     "typescript": "^5.3.2"
   },
