--- src/app/_trpc/client.ts
@@ -1,4 +1,0 @@
-import { createTRPCReact } from "@trpc/react-query";
-import { type AppRouter } from "@/server/routers";
-
-export const trpc = createTRPCReact<AppRouter>({ abortOnUnmount: true });
