--- src/app/auth/layout.tsx
@@ -1,5 +1,0 @@
-import React, { PropsWithChildren } from "react";
-
-export default function AuthLayout({ children }: PropsWithChildren) {
-	return <div className="min-h-screen">{children}</div>;
-}
