--- src/app/auth/login/page.tsx
@@ -1,53 +1,0 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-"use client";
-
-import Image from "next/image";
-import React from "react";
-
-export default function LoginPage() {
-	const loginUser = async () => ({});
-
-	return (
-		<div className="flex h-screen w-screen flex-col items-center justify-center">
-			<div className="mt-16 grid space-y-4">
-				<button
-					className="group h-12 rounded-full border-2 border-gray-300 px-6 transition duration-300
- hover:border-blue-400 focus:bg-blue-50 active:bg-blue-100"
-					onClick={() => loginUser()}
-				>
-					<div className="relative flex items-center justify-center space-x-2">
-						<Image
-							src="https://assets-global.website-files.com/6257adef93867e50d84d30e2/636e0a6a49cf127bf92de1e2_icon_clyde_blurple_RGB.png"
-							className="w-5 text-sky-700"
-							width={20}
-							height={20}
-							alt="google logo"
-						/>
-						<span className="block w-max text-sm font-semibold tracking-wide text-gray-700 transition duration-300 group-hover:text-blue-600 sm:text-base">
-							Login with Discord
-						</span>
-					</div>
-				</button>
-				<button
-					className="group h-12 rounded-full border-2 border-gray-300 px-6 transition duration-300
- hover:border-blue-400 focus:bg-blue-50 active:bg-blue-100"
-					onClick={() => loginUser()}
-				>
-					<div className="relative flex items-center justify-center space-x-2">
-						<svg
-							xmlns="http://www.w3.org/2000/svg"
-							fill="currentColor"
-							className="w-5 text-gray-700"
-							viewBox="0 0 16 16"
-						>
-							<path d="M8 0C3.58 0 0 3.58 0 8c0 3.54 2.29 6.53 5.47 7.59.4.07.55-.17.55-.38 0-.19-.01-.82-.01-1.49-2.01.37-2.53-.49-2.69-.94-.09-.23-.48-.94-.82-1.13-.28-.15-.68-.52-.01-.53.63-.01 1.08.58 1.23.82.72 1.21 1.87.87 2.33.66.07-.52.28-.87.51-1.07-1.78-.2-3.64-.89-3.64-3.95 0-.87.31-1.59.82-2.15-.08-.2-.36-1.02.08-2.12 0 0 .67-.21 2.2.82.64-.18 1.32-.27 2-.27.68 0 1.36.09 2 .27 1.53-1.04 2.2-.82 2.2-.82.44 1.1.16 1.92.08 2.12.51.56.82 1.27.82 2.15 0 3.07-1.87 3.75-3.65 3.95.29.25.54.73.54 1.48 0 1.07-.01 1.93-.01 2.2 0 .21.15.46.55.38A8.012 8.012 0 0 0 16 8c0-4.42-3.58-8-8-8z" />
-						</svg>
-						<span className="block w-max text-sm font-semibold tracking-wide text-gray-700 transition duration-300 group-hover:text-blue-600 sm:text-base">
-							Login with Github
-						</span>
-					</div>
-				</button>
-			</div>
-		</div>
-	);
-}
