--- src/app/auth/signup/page.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export default function SignupPage() {
-	return (
-		<div className="container flex h-screen w-screen flex-col items-center justify-center">
-			This is the Signup page.
-		</div>
-	);
-}
