--- src/components/ui/toaster.tsx
@@ -1,35 +1,0 @@
-"use client";
-
-import { useToast } from "@/components/ui/use-toast";
-import {
-	ToastDescription,
-	ToastProvider,
-	ToastViewport,
-	ToastTitle,
-	ToastClose,
-	Toast,
-} from "@/components/ui/toast";
-
-export function Toaster() {
-	const { toasts } = useToast();
-
-	return (
-		<ToastProvider>
-			{toasts.map(function ({ id, title, description, action, ...props }) {
-				return (
-					<Toast key={id} {...props}>
-						<div className="grid gap-1">
-							{title && <ToastTitle>{title}</ToastTitle>}
-							{description && (
-								<ToastDescription>{description}</ToastDescription>
-							)}
-						</div>
-						{action}
-						<ToastClose />
-					</Toast>
-				);
-			})}
-			<ToastViewport />
-		</ToastProvider>
-	);
-}
