--- src/tests/unit/index.test.ts
@@ -1,7 +1,0 @@
-import { expect, test } from "@playwright/test";
-
-test.setTimeout(35e3);
-
-test("Sample test", () => {
-	expect(true).toBeTruthy();
-});
